// MyApp/Angular12Crud/src/app/models/job-application.model.ts
export class Job_application {
  id?: any;
  id_jobseeker?: any;
  id_hr?: any;
  id_advertisement?: any;
  jobseeker_messages?: string;
  advertiser_messages?: string;
  status?: string;
  date?: string;
  age?: any;
  work_experience?: string;
  education?: string;
  first_name?: string;
  last_name?: string;
  email?: string;
  phone?: any;
}